requirements.txt
Django~=5.0.7
automium_web~=0.1.1
environs~=9.3.0
pytelegrambotapi==4.20.0
phonenumbers==8.13.40
django-phonenumber-field==8.0.0
schedule==1.2.2